chore(ExploreProjectsPage): rename files
