+++ src/pages/explore/top.tsx
@@ -1,4 +1,4 @@
-import { ExploreProjectsPage } from '../../components/ExporeProjectsPage/ExporeProjectsPage';
+import { ExploreProjectsPage } from '../../components/ExploreProjectsPage/ExploreProjectsPage';
 import { declareSsrProps } from '../../utils/declareSsrProps';
 
 export const getServerSideProps = declareSsrProps(
